+++ addons/crm/base_partner_merge.py
@@ -227,7 +227,7 @@ class MergePartnerAutomatic(osv.TransientModel):
             update_records('ir.attachment', src=partner, field_model='res_model')
             update_records('mail.followers', src=partner, field_model='res_model')
             update_records('mail.message', src=partner)
-            update_records('marketing.campaign.workitem', src=partner, field_model='object_id.model')
+            #update_records('marketing.campaign.workitem', src=partner, field_model='object_id.model')
             update_records('ir.model.data', src=partner)
 
         proxy = self.pool['ir.model.fields']