+++ addons/crm/crm.py
@@ -35,8 +35,8 @@ AVAILABLE_PRIORITIES = [
     ('4', 'Very High'),
 ]
 
-class crm_case_channel(osv.osv):
-    _name = "crm.case.channel"
+class crm_channel(osv.osv):
+    _name = "crm.channel"
     _description = "Channels"
     _order = 'name'
     _columns = {
@@ -47,13 +47,13 @@ class crm_case_channel(osv.osv):
         'active': lambda *a: 1,
     }
 
-class crm_case_stage(osv.osv):
+class crm_stage(osv.osv):
     """ Model for case stages. This models the main stages of a document
         management flow. Main CRM objects (leads, opportunities, project
         issues, ...) will now use only stages, instead of state and stages.
         Stages are for example used to display the kanban view of records.
     """
-    _name = "crm.case.stage"
+    _name = "crm.stage"
     _description = "Stage of case"
     _rec_name = 'name'
     _order = "sequence"
@@ -64,7 +64,7 @@ class crm_case_stage(osv.osv):
         'probability': fields.float('Probability (%)', required=True, help="This percentage depicts the default/average probability of the Case for this stage to be a success"),
         'on_change': fields.boolean('Change Probability Automatically', help="Setting this stage will change the probability automatically on the opportunity."),
         'requirements': fields.text('Requirements'),
-        'section_ids': fields.many2many('crm.case.section', 'section_stage_rel', 'stage_id', 'section_id', string='Sections',
+        'section_ids': fields.many2many('crm.team', 'section_stage_rel', 'stage_id', 'section_id', string='Sections',
                         help="Link between stages and sales teams. When set, this limitate the current stage to the selected sales teams."),
         'case_default': fields.boolean('Default to New Sales Team',
                         help="If you check this field, this stage will be proposed by default on each sales team. It will not assign this stage to existing teams."),
@@ -86,44 +86,53 @@ class crm_case_stage(osv.osv):
         'type': 'both',
         'case_default': True,
     }
-
-class crm_case_categ(osv.osv):
+class crm_case_category(osv.Model):
     """ Category of Case """
-    _name = "crm.case.categ"
+    _name = "crm.case.category"
     _description = "Category of Case"
     _columns = {
         'name': fields.char('Name', required=True, translate=True),
-        'section_id': fields.many2one('crm.case.section', 'Sales Team'),
-        'object_id': fields.many2one('ir.model', 'Object Name'),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
     }
-    def _find_object_id(self, cr, uid, context=None):
-        """Finds id for case object"""
-        context = context or {}
-        object_id = context.get('object_id', False)
-        ids = self.pool.get('ir.model').search(cr, uid, ['|',('id', '=', object_id),('model', '=', context.get('object_name', False))])
-        return ids and ids[0] or False
-    _defaults = {
-        'object_id' : _find_object_id
+class crm_phonecall_category(osv.Model):
+    """ Category of Case """
+    _name = "crm.phonecall.category"
+    _description = "Category of phonecall"
+    _columns = {
+        'name': fields.char('Name', required=True, translate=True),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
     }
-
-class crm_case_resource_type(osv.osv):
+class crm_lead_tag(osv.Model):
+    """ Category of Case """
+    _name = "crm.lead.tag"
+    _description = "Category of lead"
+    _columns = {
+        'name': fields.char('Name', required=True, translate=True),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
+    }
+class crm_claim_category(osv.Model):
+    """ Category of Case """
+    _name = "crm.claim.category"
+    _description = "Category of claim"
+    _columns = {
+        'name': fields.char('Name', required=True, translate=True),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
+    }
+class crm_helpdesk_category(osv.Model):
+    """ Category of Case """
+    _name = "crm.helpdesk.category"
+    _description = "Category of helpdesk"
+    _columns = {
+        'name': fields.char('Name', required=True, translate=True),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
+    }
+class crm_campaign(osv.osv):
     """ Resource Type of case """
-    _name = "crm.case.resource.type"
+    _name = "crm.campaign"
     _description = "Campaign"
     _rec_name = "name"
     _columns = {
         'name': fields.char('Campaign Name', required=True, translate=True),
-        'section_id': fields.many2one('crm.case.section', 'Sales Team'),
+        'section_id': fields.many2one('crm.team', 'Sales Team'),
     }
-
-class crm_payment_mode(osv.osv):
-    """ Payment Mode for Fund """
-    _name = "crm.payment.mode"
-    _description = "CRM Payment Mode"
-    _columns = {
-        'name': fields.char('Name', required=True),
-        'section_id': fields.many2one('crm.case.section', 'Sales Team'),
-    }
-
-
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4: