+++ addons/crm/sales_team.py
@@ -9,8 +9,8 @@ from openerp import tools
 from openerp.osv import fields, osv
 
 
-class crm_case_section(osv.Model):
-    _inherit = 'crm.case.section'
+class crm_team(osv.Model):
+    _inherit = 'crm.team'
     _inherits = {'mail.alias': 'alias_id'}
 
     def _get_opportunities_data(self, cr, uid, ids, field_name, arg, context=None):
@@ -39,7 +39,7 @@ class crm_case_section(osv.Model):
 
     _columns = {
         'resource_calendar_id': fields.many2one('resource.calendar', "Working Time", help="Used to compute open days"),
-        'stage_ids': fields.many2many('crm.case.stage', 'section_stage_rel', 'section_id', 'stage_id', 'Stages'),
+        'stage_ids': fields.many2many('crm.stage', 'section_stage_rel', 'section_id', 'stage_id', 'Stages'),
         'use_leads': fields.boolean('Leads',
             help="The first contact you get with a potential customer is a lead you qualify before converting it into a real business opportunity. Check this box to manage leads in this sales team."),
         'use_opportunities': fields.boolean('Opportunities', help="Check this box to manage opportunities in this sales team."),
@@ -54,11 +54,11 @@ class crm_case_section(osv.Model):
 
     def _auto_init(self, cr, context=None):
         """Installation hook to create aliases for all lead and avoid constraint errors."""
-        return self.pool.get('mail.alias').migrate_to_alias(cr, self._name, self._table, super(crm_case_section, self)._auto_init,
+        return self.pool.get('mail.alias').migrate_to_alias(cr, self._name, self._table, super(crm_team, self)._auto_init,
             'crm.lead', self._columns['alias_id'], 'name', alias_prefix='Lead+', alias_defaults={}, context=context)
 
     def _get_stage_common(self, cr, uid, context):
-        ids = self.pool.get('crm.case.stage').search(cr, uid, [('case_default', '=', 1)], context=context)
+        ids = self.pool.get('crm.stage').search(cr, uid, [('case_default', '=', 1)], context=context)
         return ids
 
     _defaults = {
@@ -71,7 +71,7 @@ class crm_case_section(osv.Model):
         if context is None:
             context = {}
         create_context = dict(context, alias_model_name='crm.lead', alias_parent_model_name=self._name)
-        section_id = super(crm_case_section, self).create(cr, uid, vals, context=create_context)
+        section_id = super(crm_team, self).create(cr, uid, vals, context=create_context)
         section = self.browse(cr, uid, section_id, context=context)
         self.pool.get('mail.alias').write(cr, uid, [section.alias_id.id], {'alias_parent_thread_id': section_id, 'alias_defaults': {'section_id': section_id, 'type': 'lead'}}, context=context)
         return section_id
@@ -80,6 +80,6 @@ class crm_case_section(osv.Model):
         # Cascade-delete mail aliases as well, as they should not exist without the sales team.
         mail_alias = self.pool.get('mail.alias')
         alias_ids = [team.alias_id.id for team in self.browse(cr, uid, ids, context=context) if team.alias_id]
-        res = super(crm_case_section, self).unlink(cr, uid, ids, context=context)
+        res = super(crm_team, self).unlink(cr, uid, ids, context=context)
         mail_alias.unlink(cr, uid, alias_ids, context=context)
         return res