+++ addons/crm_helpdesk/crm_helpdesk.py
@@ -46,7 +46,7 @@ class crm_helpdesk(osv.osv):
             'write_date': fields.datetime('Update Date' , readonly=True),
             'date_deadline': fields.date('Deadline'),
             'user_id': fields.many2one('res.users', 'Responsible'),
-            'section_id': fields.many2one('crm.case.section', 'Sales Team', \
+            'section_id': fields.many2one('crm.team', 'Sales Team', \
                             select=True, help='Responsible sales team. Define Responsible user and Email account for mail gateway.'),
             'company_id': fields.many2one('res.company', 'Company'),
             'date_closed': fields.datetime('Closed', readonly=True),
@@ -56,14 +56,13 @@ class crm_helpdesk(osv.osv):
             'date': fields.datetime('Date'),
             'ref': fields.reference('Reference', selection=openerp.addons.base.res.res_request.referencable_models),
             'ref2': fields.reference('Reference 2', selection=openerp.addons.base.res.res_request.referencable_models),
-            'channel_id': fields.many2one('crm.case.channel', 'Channel', help="Communication channel."),
+            'channel_id': fields.many2one('crm.channel', 'Channel', help="Communication channel."),
             'planned_revenue': fields.float('Planned Revenue'),
             'planned_cost': fields.float('Planned Costs'),
             'priority': fields.selection([('0','Low'), ('1','Normal'), ('2','High')], 'Priority'),
             'probability': fields.float('Probability (%)'),
-            'categ_id': fields.many2one('crm.case.categ', 'Category', \
-                            domain="['|',('section_id','=',False),('section_id','=',section_id),\
-                            ('object_id.model', '=', 'crm.helpdesk')]"),
+            'categ_id': fields.many2one('crm.helpdesk.category', 'Category', \
+                            domain="['|',('section_id','=',False),('section_id','=',section_id)]"),
             'duration': fields.float('Duration', states={'done': [('readonly', True)]}),
             'state': fields.selection(
                 [('draft', 'New'),