+++ addons/sale_crm/sale_crm.py
@@ -24,6 +24,6 @@ from openerp.osv import osv, fields
 class sale_order(osv.osv):
     _inherit = 'sale.order'
     _columns = {
-        'categ_ids': fields.many2many('crm.case.categ', 'sale_order_category_rel', 'order_id', 'category_id', 'Tags', \
-            domain="['|', ('section_id', '=', section_id), ('section_id', '=', False), ('object_id.model', '=', 'crm.lead')]", context="{'object_name': 'crm.lead'}")
+        'tag_ids': fields.many2many('crm.lead.tag', 'sale_order_category_rel', 'order_id', 'category_id', 'Tags', \
+            domain="['|', ('section_id', '=', section_id), ('section_id', '=', False)]",context="{'object_name': 'crm.lead'}")
     }